src/session/requirements.txt
json5==0.9.25
requests==2.32.3
numpy==1.26.4
mediapipe==0.10.20
opencv-python==4.10.0.84
pydantic==2.10.5
pydantic_core==2.27.2
pyserial==3.5
matplotlib==3.10.0